feat: Add version 3 of main.py.
Significant gains made here with the usage of generator comprehension instead of list comprehension. We also cleverly made join() the consumer of that generator. Issues with memory arise from sorted() however. Sorting will likely need to be tackled when we are gathering the lines of data from the file (in the code directly above). Lastly it seems writing the code in a way that requires less memory read and writes, and freeing up more memory for the program are key parts to optimizing this. Another thing to try is parallel computation with concurrent futures and Process are avenues to explore.